Swapping now works. Implemented different schduling methods, and shell command. Altered load and loadall to except priority. Priority scheduling has errors I need to fix. Other than that, I need to: 1. Have queus list priority, make UI look pretty, and test, test, test.

+++ TSOS/source/os/memorymanager.ts
@@ -242,6 +242,17 @@ module TSOS
             return dword;
         }
 
+        //          CR_FILE_NOT_FOUND on file not found,
+        //          CR_DRIVE_NOT_FORMATED if drive not formatted,
+        //          CR_DID_NOT_WRITE_ALL_DATA  if text uses more space than available.
+        //          CR_CORRUPTED_FILE_BLOCK if missing a file block
+
+        //          CR_FILE_LENGTH_TO_LONG if file name to long,
+        //          CR_EMPTY_FILE_NAME if empty file name,
+        //          CR_FILE_DIRECTORY_FULL  if no more space in file directory,
+        //          CR_DRIVE_FULL if no free blocks.
+
+        //          CR_FILE_NOT_FOUND if file not in directory
         public loadToHDDBytes(data : number[],pid:number) : any[]
         {
             // Inits
@@ -249,7 +260,7 @@ module TSOS
             var ret : number;
             var error : boolean = false;
             var val : string;
-            var data : number[] = [];
+            var wdata : number[] = [];
 
             // Create swap file name
             fname = "~proc-" + pid.toString();
@@ -282,14 +293,15 @@ module TSOS
                 {
                     if( i < data.length)
                     {
-                        data.push(data[i]);
+                        wdata.push(data[i]);
                     }
                     else
-                        data.push(0);
+                        wdata.push(0);
                 }
 
+
                 // Write to file
-                ret = _HDDriver.writeToFileBytes(fname,data);
+                ret = _HDDriver.writeToFileBytes(fname,wdata);
 
                 // If not error, put pid on loaded hdd pid array
                 if( ret == CR_SUCCESS)
@@ -476,32 +488,41 @@ module TSOS
             var oldfname : string;
             var oldpid : number;
             var swappid : number;
+            var savePartToHD : boolean = false;
 
             // Check if valid partition
             if( partition < 0 || partition >= _MemoryPartitions)
-                return -1;
+                return CR_INVALID_PARTITION;
 
             // Check if hdpid is currently on hd
             if( this.findHDLoadedPID(hdpid) == -1 )
-                return -1;
+                return CR_SWAP_FILE_NOT_LOADED;
+
+            // Check if partition is free
+            if( this.partitionsLoaded[partition] )
+            {
+
+                // Get partition bytes
+                data = this.getPartitionBytes(partition);
+
+                // Create swap file
+                aret = this.loadToHDDBytes(data, this.partitionPIDs[partition]);
 
 
-            // Get partition bytes
-            data = this.getPartitionBytes(partition);
-
-            // Create swap file
-            aret = this.loadToHDDBytes(data,this.partitionPIDs[partition]);
+                // Return error if could not load
+                if (aret[0] < 0)
+                    return aret[0];
 
 
+                // Get swap fname if error occurs and need to delete
+                oldfname = aret[1];
+                oldpid = this.partitionPIDs[partition];
 
-            // Return error if could not load
-            if( aret[0] < 0)
-                return -1;
-
-
-            // Get swap fname if error occurs and need to delete
-            oldfname = aret[1];
-            oldpid = this.partitionPIDs[partition];
+                // Set save to hd flag
+                savePartToHD = true;
+            }
+            else
+                oldpid = CR_MEMORY_SWAP_FILE_NOT_NEEDED;
 
             // Get hd swap file bytes
             aret = _HDDriver.readFileBytes(this.findHDLoadedFName(hdpid));
@@ -509,11 +530,12 @@ module TSOS
             // Verify valid data
             if( aret[0] < 0)
             {
-                // Remove created swap file
-                _HDDriver.deleteFile(oldfname);
+                // Remove created swap file if one created
+                if( savePartToHD )
+                     _HDDriver.deleteFile(oldfname);
 
                 // Return error
-                return -1;
+                return aret[0];
             }
 
 
@@ -521,12 +543,15 @@ module TSOS
             if( !this.loadPartitionBytes(partition,aret[1]) )
             {
                 // Remove created swap file
-                _HDDriver.deleteFile(oldfname);
+                if( savePartToHD )
+                    _HDDriver.deleteFile(oldfname);
 
                 // Return error
-                return -1;
+                return CR_PARTITION_NOT_LOADED;
             }
 
+            // Set partition in use
+            this.partitionsLoaded[partition] = true;
 
             // Change loaded pid
             this.partitionPIDs[partition] = hdpid;
